+++ src/utils/AppRouter.tsx
@@ -1,11 +1,12 @@
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import HomePage from '../pages/home';
+import MoviePage from '../pages/movie';
 const AppRouter = () => {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<HomePage/>} />
-        <Route path="/movie/:id" element={<div>Movie</div>} />
+        <Route path="/movie/:id" element={<MoviePage/>} />
         <Route path="*" element={<Navigate to="/" />} />
       </Routes>
     </Router>